Move MergePoint from compile.rs to trace_ctx.rs for RPython parity
RPython's MetaInterp (pyjitpl.py) owns both the MergePoint tuple and
the current_merge_points list. Move the struct definition to live
alongside its owning field in trace_ctx.rs rather than compile.rs.
